+++ src/Core/Tables/Columns/DateColumn.php
@@ -1,0 +1,20 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Pepperfm\Flashboard\Core\Tables\Columns;
+
+class DateColumn extends Column
+{
+    public const string TYPE = 'date';
+
+    public static function make(string $key): static
+    {
+        return parent::make($key)->type(self::TYPE);
+    }
+
+    public function format(string $format): static
+    {
+        return $this->attribute('format', $format);
+    }
+}